# Extension perf improvments