@Telroshan 2024-07-12
Hey, while I understand the theoretical performance concern, did you measure the actual performance impact on say a large DOM hierarchy or, more broadly, the worst case scenario? To see what the impact is in this case. 

I see 2 downsides to the suggested approach:
- I don't know all existing extensions but what if an extension has "dynamic selectors" of some sort? What would have been evaluated at every call until now would suddenly be evaluated only at the extension register moment, I'm afraid it could be a breaking change
- It adds more hardcoded strings in the code, which cannot be minified, and we'd like to avoid that as much as possible as it increases the lib size. A minor inconvenience here given the amount of text, I'll agree, but if the performance impact turns out to be even more minor than that... Hence the measurement request!

Note that imho there's no need to address any of the 2 points mentioned above, before a performance test case is produced.
I'd like to first see codesamples and their performance result, both for the current lib state, and the suggested one, before moving on with the retro-compatibility / minification issues!
As after all, if the performance gain is interesting, we'll find a way to work around those constraints anyway.