@MichaelWest22 2024-07-12
thanks for looking at this @Telroshan

With regard to the first possible downside I researched when this feature was added and it was very recently only a few months ago in #2505 where it was added and as far as I can see from all my searches only so far used in SSE extension to solve a duplicate processing issue and seems to be only have been intended for a static selector list which is why I thought it could be optimized.

But yeah I had not considered the final minified gziped size.  should probably run that quickly to check the change.

Will see if I can work on some perf tests.  With small virtually empty DOM like all my testing so far the performance difference would be impossible to notice I think so i may have to get a more realistic page and repeat the tasks multiple times.