@MichaelWest22 2024-07-12
```
Before
-rwxrwxrwx 1 mwest mwest  49082 Jul 12 20:30 htmx.min.js
-rwxrwxrwx 1 mwest mwest  15761 Jul 12 20:30 htmx.min.js.gz
After
-rwxrwxrwx 1 mwest mwest  49267 Jul 12 20:28 htmx.min.js     + 185
-rwxrwxrwx 1 mwest mwest  15812 Jul 12 20:28 htmx.min.js.gz  + 51
``` 
So it does add a little to the min size as you suggested

Perf testing this change is not 100% easy as the calls i'm optimizing are only a small % of the total time to do real requests so maybe this perf optimization is just not really needed.

One method i've found to get something meaningful so far is to add the following code  before getExtension and rename getExtension to getExtension2
```
  function getExtensions(elt) {
    var ret
    for (var i = 1; i <= 1000; i++) {
      ret = getExtensions2(elt)
    }
    return ret
  }
```
What this does is makes the extension lookup code ~1000 times worse so I can isolate just this portion of the code.  Then i run the full htmx test suite except with the 3 extension test scripts commented out.

I got the following results:

```
Before my perf Optimizations:
    ✅ htmx:sendError is called after a failed request (40ms)
    ✅ htmx:afterRequest is called when replacing outerHTML (31ms)
    ✅ htmx:afterOnLoad is called when replacing outerHTML (19ms)
    ✅ htmx:beforeProcessNode is called when replacing outerHTML (20ms)
    ✅ htmx:beforeProcessNode allows htmx attribute tweaking (23ms)
    ✅ htmx:afterProcessNode is called after replacing outerHTML (29ms)
    ✅ htmx:afterRequest is called when targeting a parent div (30ms)
    ✅ adding an error in htmx:configRequest stops the request (5ms)
    ✅ preventDefault() in htmx:configRequest stops the request (5ms)
    ✅ preventDefault() in the htmx:beforeRequest event cancels the request (5ms)
    ✅ preventDefault() in the htmx:beforeOnLoad event cancels the swap (20ms)
    ✅ htmx:afterRequest event contains 'successful' and 'failed' properties indicating success after successful request (18ms)
Response Status Error Code 500 from /test
    ✅ htmx:afterRequest event contains 'successful' and 'failed' properties indicating failure after failed request (25ms)
    ✅ htmx:confirm can cancel request (29ms)
beforeSwap {"htmx-internal-data":{"initHash":3227304}} {"htmx-internal-data":{"initHash":3227305}}
afterSwap {"htmx-internal-data":{"initHash":3227305}} {"htmx-internal-data":{"initHash":3227305}}
    ✅ has updated target available when target set via htmx:beforeSwap (21ms)


  632 passing (19s)
  3 pending
```

```
After my perf Optimizations:
    ✅ htmx:sendError is called after a failed request (32ms)
    ✅ htmx:afterRequest is called when replacing outerHTML (1ms)
    ✅ htmx:afterOnLoad is called when replacing outerHTML (1ms)
    ✅ htmx:beforeProcessNode is called when replacing outerHTML (1ms)
    ✅ htmx:beforeProcessNode allows htmx attribute tweaking (1ms)
    ✅ htmx:afterProcessNode is called after replacing outerHTML (1ms)
    ✅ htmx:afterRequest is called when targeting a parent div (1ms)
    ✅ adding an error in htmx:configRequest stops the request
    ✅ preventDefault() in htmx:configRequest stops the request (1ms)
    ✅ preventDefault() in the htmx:beforeRequest event cancels the request
    ✅ preventDefault() in the htmx:beforeOnLoad event cancels the swap (1ms)
    ✅ htmx:afterRequest event contains 'successful' and 'failed' properties indicating success after successful request (1ms)
Response Status Error Code 500 from /test
    ✅ htmx:afterRequest event contains 'successful' and 'failed' properties indicating failure after failed request (1ms)
    ✅ htmx:confirm can cancel request (1ms)
beforeSwap {"htmx-internal-data":{"initHash":3227304}} {"htmx-internal-data":{"initHash":3227305}}
afterSwap {"htmx-internal-data":{"initHash":3227305}} {"htmx-internal-data":{"initHash":3227305}}
    ✅ has updated target available when target set via htmx:beforeSwap (1ms)


  632 passing (3s)
  3 pending
```
So from this quick test it shows that it saved 16 seconds compared to the 3 second normal run.  Both runs had the 1000 times extension checking loop code but with my updated version this didn't add any time to the normal 3 second test suite.  If we divide my very impresive 16 second speed gain by 1000 to find the real perf gain and we are then down to just 16ms which does not seem that impressive you can say I just saved 16 ms in a 3000ms test suite which is like 0.5% perf gain for a 0.32% bigger gz size.  

Also when I include the 3 extension tests it didn't change the 19 second old code run timing by much but my code then took 13 seconds and was much slower than before.  This is because once you get 1/4 of the way though the tests and it loads the extensions most if not all of the extension actions get activated and this disables my optimizations as it only gives perf savings when only some extension actions are enabled.

So I think it does give a good speed gain for the situations where extensions are used sparingly but it is all sub ms savings per request so may not be worth the effort.  May still be useful to generate a more realistic complex web page and do some more testing on that though.